+++ src/main/java/com/ll/demo/comment/dto/comment/CreateCommentRequest.java
@@ -1,4 +1,4 @@
-package com.ll.demo.comment.dto;
+package com.ll.demo.comment.dto.comment;
 
 import com.ll.demo.article.entity.Article;
 import com.ll.demo.comment.entity.Comment;
@@ -13,23 +13,18 @@ import org.springframework.validation.annotation.Validated;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
 @Validated
-public class UpdateCommentRequest {
-    @Setter
-    private Long commentId;
-
+public class CreateCommentRequest {
     @NotNull(message = "어떤 게시글의 아이디인지 확인이 필요합니다.")
     private Long articleId;
 
     @NotNull(message = "어떤 멤버인지 확인이 필요합니다.")
     private String username;
 
-    @Setter
     @NotEmpty(message = "댓글 내용은 필수입니다.")
     private String content;
 
-    public static Comment toEntity(UpdateCommentRequest request, Member member, Article article) {
+    public static Comment toEntity(CreateCommentRequest request, Member member, Article article) {
         return Comment.builder()
-                .id(request.getCommentId())
                 .content(request.getContent())
                 .member(member)
                 .article(article)
