+++ src/main/java/com/ll/demo/comment/dto/comment/CreateCommentResponse.java
@@ -1,4 +1,4 @@
-package com.ll.demo.comment.dto;
+package com.ll.demo.comment.dto.comment;
 
 import com.ll.demo.comment.entity.Comment;
 import lombok.*;
@@ -7,13 +7,13 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
-public class UpdateReplyCommentResponse {
-    private Long replyId;
+public class CreateCommentResponse {
+    private Long id;
     private String content;
 
-    public static UpdateReplyCommentResponse of(Comment comment) {
-        return UpdateReplyCommentResponse.builder()
-                .replyId(comment.getId())
+    public static CreateCommentResponse of(Comment comment) {
+        return CreateCommentResponse.builder()
+                .id(comment.getId())
                 .content(comment.getContent())
                 .build();
     }
