+++ src/main/java/com/ll/demo/comment/dto/comment/DeleteCommentRequest.java
@@ -1,9 +1,8 @@
-package com.ll.demo.comment.dto;
+package com.ll.demo.comment.dto.comment;
 
 import com.ll.demo.article.entity.Article;
 import com.ll.demo.comment.entity.Comment;
 import com.ll.demo.member.entity.Member;
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import org.springframework.validation.annotation.Validated;
@@ -13,19 +12,19 @@ import org.springframework.validation.annotation.Validated;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
 @Validated
-public class CreateCommentRequest {
+public class DeleteCommentRequest {
+    @Setter
+    private Long commentId;
+
     @NotNull(message = "어떤 게시글의 아이디인지 확인이 필요합니다.")
     private Long articleId;
 
     @NotNull(message = "어떤 멤버인지 확인이 필요합니다.")
     private String username;
 
-    @NotEmpty(message = "댓글 내용은 필수입니다.")
-    private String content;
-
-    public static Comment toEntity(CreateCommentRequest request, Member member, Article article) {
+    public static Comment toEntity(DeleteCommentRequest request, Member member, Article article) {
         return Comment.builder()
-                .content(request.getContent())
+                .id(request.getCommentId())
                 .member(member)
                 .article(article)
                 .build();
