+++ src/main/java/com/ll/demo/comment/dto/reply/DeleteReplyCommentResponse.java
@@ -1,4 +1,4 @@
-package com.ll.demo.comment.dto;
+package com.ll.demo.comment.dto.reply;
 
 import com.ll.demo.comment.entity.Comment;
 import lombok.*;
@@ -6,15 +6,19 @@ import lombok.*;
 import java.time.LocalDateTime;
 
 @Getter
+@Setter
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
-public class DeleteCommentResponse {
+public class DeleteReplyCommentResponse {
     private Long replyId;
+    private String content;
     private LocalDateTime removedTime;
-    public static DeleteCommentResponse of(Comment comment) {
-        return DeleteCommentResponse.builder()
+
+    public static DeleteReplyCommentResponse of(Comment comment) {
+        return DeleteReplyCommentResponse.builder()
                 .replyId(comment.getId())
+                .content(comment.getContent())
                 .removedTime(comment.getRemovedTime())
                 .build();
     }
